requirements.txt
python-dotenv~=1.0.1
requests~=2.32.4
pandas~=2.3.0
praw~=7.8.1
torch~=2.7.1
transformers~=4.53.0